Half-Life Simulator v1.0
The radiation symbol gif is required for this program to work.  This program simulates half-lives of a number of radioactive isotopes, telling you the mass of each after however many half-lives pass.  The user selects a radioactive isotope, then clicks the "Half-life" button to cause radioactive decay to happen.  The program then tells how much mass of the selected isotope is left after the number of half-lives simulated.  This program uses the Tkinter library of widgets to create a GUI look.  Most of the code uses Tkinter functions to work, so a user who downloads this program will have to have Python installed with Tkinter.