Battleship game v2.0
This is the same Battleship game with longer battleships and a different one-player mode.  Both modes involve trying to sink the other player's battleships by guessing all the appropriate squares the battleships occupy.  All 4 squares a battleship occupies must be guessed correctly for the battleship to be sunken.  The player to sink all the other player's battleships first wins.  The CPU has somewhat of an AI in one-player mode, but only randomly guesses squares every turn.  The boards may be at least 8 squares big, and each player may have at most 1/8 of the board's size worth of battleships on the board every game.  There will be revisions to the game with the CPU having smarter AI, thus introducing difficulty levels.  Later revisions may also allow each player to choose the number of battleships they want on the board for themselves and the CPU, as in the first game.