Battleship game v2.5
This is a revision of my second battleship game, with added AI for the CPU in the one-player mode allowing for three difficulty settings.  The difficulty settings are easy, intermediate and hard, where the CPU is more perceptive the harder the difficulty.  This file also requires the battleship_ii.txt file to load.  It also has revisions done with the code, like a new function for the CPU to guess a row and column under control by a difficulty decorator that modifies the function based on the difficulty chosen, and rewriting some lines to make the code more efficient.